SIG-304 Se agregan permisos al boton eliminar, se agregan mensajes de confirmacion al eliminar

+++ src/app/modulos/scm/components/formulario-scm/formulario-scm.component.ts
@@ -519,6 +519,12 @@ export class FormularioScmComponent implements OnInit, OnDestroy {
 
     }
 
+    onClick(){
+        console.log(this.sesionService.getPermisosMap())
+        console.log(this.sesionService.getPermisosMap()["SCM_DEL_CASE_DIAG"])
+
+    }
+
     async aonRegisters() {
         try {
             let token: any = await this.scmService.getTokenAon();
@@ -1029,15 +1035,24 @@ export class FormularioScmComponent implements OnInit, OnDestroy {
     async deleteDiagnostico(id) {
         this.msgs = [];
         try {
-            let resp = await this.scmService.deleteDiagnosticos(id);
-            if (resp) {
-                this.msgs.push({
-                    severity: "error",
-                    summary: "Diagnostico",
-                    detail: `Su Diagnostico fue eliminado`,
-                });
-                this.onCloseModalDianostico();
+            if (await this.confirmService.confirmDiagnostico()){
+                let resp = await this.scmService.deleteDiagnosticos(id);
+                if (resp) {
+                    this.msgs.push({
+                        severity: "error",
+                        summary: "Diagnostico",
+                        detail: `Su Diagnostico fue eliminado`,
+                    });
+                    this.onCloseModalDianostico();
+                }
+            }            
+            else {
+                this.msgs = [
+                    { severity: "info", summary: "Cancelado", detail: "usted cancelo la eliminación" }
+                ];
             }
+
+           
 
         } catch (error) {
             console.log(error)
@@ -1051,15 +1066,23 @@ export class FormularioScmComponent implements OnInit, OnDestroy {
     async deleteRecomendation(id) {
         this.msgs = [];
         try {
-            let resp = await this.scmService.deleteRecomendation(id);
-            if (resp) {
-                this.msgs.push({
-                    severity: "error",
-                    summary: "Mensaje del sistema",
-                    detail: `Su recomendación se eliminó exitosamente`,
-                });
-                this.onCloseModalrecomendation();
+            if (await this.confirmService.confirmRecomendacion()){
+                let resp = await this.scmService.deleteRecomendation(id);
+                if (resp) {
+                    this.msgs.push({
+                        severity: "error",
+                        summary: "Mensaje del sistema",
+                        detail: `Su recomendación se eliminó exitosamente`,
+                    });
+                    this.onCloseModalrecomendation();
+                }
             }
+        else {
+            this.msgs = [
+                { severity: "info", summary: "Cancelado", detail: "usted cancelo la eliminación" }
+            ];
+        }
+            
         } catch (error) {
             console.log(error)
         }
@@ -1068,15 +1091,23 @@ export class FormularioScmComponent implements OnInit, OnDestroy {
     async deleteSeguimiento(id) {
         this.msgs = [];
         try {
-            let resp = await this.scmService.deleteSeguimiento(id);
-            if (resp) {
-                this.msgs.push({
-                    severity: "error",
-                    summary: "Mensaje del sistema",
-                    detail: `Su seguimiento se eliminó exitosamente`,
-                });
-                this.fechaSeg()
+            if (await this.confirmService.confirmSeguimiento()){
+                let resp = await this.scmService.deleteSeguimiento(id);
+                if (resp) {
+                    this.msgs.push({
+                        severity: "error",
+                        summary: "Mensaje del sistema",
+                        detail: `Su seguimiento se eliminó exitosamente`,
+                    });
+                    this.fechaSeg()
+                }
             }
+        else {
+            this.msgs = [
+                { severity: "info", summary: "Cancelado", detail: "usted cancelo la eliminación" }
+            ];
+        }
+            
 
         } catch (error) {
             console.log(error)
